pyproject.toml
[tool.black]
line-length = 300

[tool.poetry]
name = "mycorrhizal-network"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mycorrhizal_network"}]

[tool.poetry.scripts]
network = 'mycorrhizal_network.cli:cli'

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
geoip2 = "^4.6.0"
scapy = "^2.5.0"
click = "^8.1.3"
selenium = "^4.8.2"
neo4j = "^5.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
